school-of-rebels-bot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
attrs==25.3.0
audioop-lts==0.2.1
discord==2.3.2
discord.py==2.5.2
dnspython==2.7.0
frozenlist==1.5.0
idna==3.10
motor==3.7.0
multidict==6.2.0
propcache==0.3.1
pymongo==4.11.3
python-dotenv==1.1.0
typing_extensions==4.13.0
yarl==1.18.3
